pyproject.toml
[project]
name = "agentic-trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "lancedb>=0.23.0",
    "langchain-community>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-ollama>=0.3.3",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.8",
    "polygon>=1.2.6",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
    "tavily-python>=0.7.7",
]
